*: int64 -> float64

+++ jws/claims.go
@@ -84,19 +84,19 @@ func (c Claims) Audience() ([]string, bool) {
 
 // Expiration retrieves claim "exp" per its type in
 // https://tools.ietf.org/html/rfc7519#section-4.1.4
-func (c Claims) Expiration() (int64, bool) {
+func (c Claims) Expiration() (float64, bool) {
 	return jwt.Claims(c).Expiration()
 }
 
 // NotBefore retrieves claim "nbf" per its type in
 // https://tools.ietf.org/html/rfc7519#section-4.1.5
-func (c Claims) NotBefore() (int64, bool) {
+func (c Claims) NotBefore() (float64, bool) {
 	return jwt.Claims(c).NotBefore()
 }
 
 // IssuedAt retrieves claim "iat" per its type in
 // https://tools.ietf.org/html/rfc7519#section-4.1.6
-func (c Claims) IssuedAt() (int64, bool) {
+func (c Claims) IssuedAt() (float64, bool) {
 	return jwt.Claims(c).IssuedAt()
 }
 
@@ -161,19 +161,19 @@ func (c Claims) SetAudience(audience ...string) {
 
 // SetExpiration sets claim "exp" per its type in
 // https://tools.ietf.org/html/rfc7519#section-4.1.4
-func (c Claims) SetExpiration(expiration int64) {
+func (c Claims) SetExpiration(expiration float64) {
 	jwt.Claims(c).SetExpiration(expiration)
 }
 
 // SetNotBefore sets claim "nbf" per its type in
 // https://tools.ietf.org/html/rfc7519#section-4.1.5
-func (c Claims) SetNotBefore(notBefore int64) {
+func (c Claims) SetNotBefore(notBefore float64) {
 	jwt.Claims(c).SetNotBefore(notBefore)
 }
 
 // SetIssuedAt sets claim "iat" per its type in
 // https://tools.ietf.org/html/rfc7519#section-4.1.6
-func (c Claims) SetIssuedAt(issuedAt int64) {
+func (c Claims) SetIssuedAt(issuedAt float64) {
 	jwt.Claims(c).SetIssuedAt(issuedAt)
 }
